Safari-611.1.10/LayoutTests/platform/mac-wk1/fast/scrolling/latching/iframe-latch-small-deltas-expected.txt


Tests that a second scroll with small x/y deltas uses latching from an earlier scroll.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframeTarget.contentWindow.pageYOffset is 0
PASS window.pageYOffset is 200
After scroll
FAIL iframeTarget.contentWindow.pageYOffset should be 400. Was 205.
PASS window.pageYOffset is 200
After wait
FAIL iframeTarget.contentWindow.pageYOffset should be 400. Was 205.
PASS window.pageYOffset is 200
FAIL iframeTarget.contentWindow.pageYOffset should be 380. Was 200.
PASS window.pageYOffset is 200
PASS successfullyParsed is true

TEST COMPLETE
